fix: move Math utils to math package
